Better files
I think I now created the files and the header files right. Wrestled
with the compiler a bunch, then figured out it didn’t make any sense to
just compile Files.cc, and that the mile-long errors I was getting were
because I did not have a main()  in the file I was compliling. Finally
got it to work with g++ -o permutations permutations.cc Files.cc   g++
seems easier than c++ which is the compiler I was previously using.